Create 1 - Two Sum
First, after seeing the problem, I thought of brute force approach everyone thinks. But, the time complexity of brute force approach is O(n^2). And after understanding it for a while, thought of a One-pass approach in which, there is a dictionary and the differences between the target variable and the number present is calculated. If a new number while iterated is equivalent to any key of the dictionary, its index and the respective key's value is returned at the end.